Splitting the package into voici-core and voici
