+++ .github/workflows/frontend.yml
@@ -1,5 +1,9 @@
 name: Build And Check Frontend Packages
 
+defaults:
+  run:
+    shell: bash -l {0}
+
 on:
   push:
     branches:
@@ -18,13 +22,14 @@ jobs:
       uses: actions/checkout@v3
 
     - name: Install Micromamba
-      uses: mamba-org/provision-with-micromamba@main
+      uses: mamba-org/setup-micromamba@v1
       with:
-        extra-specs: |
-          python=${{ matrix.python_version }}
+        environment-file: environment.yml
 
     - name: Lint
       run: |
         yarn
         yarn run eslint:check
         yarn run prettier:check
+      working-directory:
+        python/voici-core
