+++ .github/workflows/main.yml
@@ -29,15 +29,20 @@ jobs:
     - uses: actions/checkout@v3
 
     - name: Install Micromamba
-      uses: mamba-org/provision-with-micromamba@main
+      uses: mamba-org/setup-micromamba@v1
       with:
-        extra-specs: |
-          python=${{ matrix.python_version }}
+        environment-file: environment.yml
 
     - name: Lint (Python)
       run: |
         python -m ruff check .
         python -m black --check .
+
+    - name: Install voici
+      run: |
+        python -m pip install jupyterlite-xeus-python
+        python -m pip install python/voici-core
+        python -m pip install python/voici
 
     - name: Install browser
       run: |
@@ -47,7 +52,6 @@ jobs:
 
     - name: Install dependencies and build tests
       run: |
-        python -m pip install .
         cd ui-tests
         yarn install --immutable
         yarn run build
