+++ environment.yml
@@ -4,6 +4,7 @@ channels:
 dependencies:
   - pip
   - black
+  - ruff
   - pytest
   - yarn=3
   - nodejs=18
@@ -12,4 +13,3 @@ dependencies:
     - voila>=0.5.5
     - voila-material>=0.4.3
     - jupyterlite-core[lab]>=0.2.0,<0.3.0
-    - jupyterlite-xeus-python >= 1.0.0a3
