+++ python/voici-core/.eslintignore
