+++ python/voici-core/.prettierignore
