+++ python/voici-core/.prettierrc
