+++ python/voici-core/LICENSE
