+++ python/voici-core/packages/voici/src/global.d.ts
