+++ python/voici-core/packages/voici/webpack.config.js
@@ -52,7 +52,7 @@ const distRoot = path.resolve(
   __dirname,
   '..',
   '..',
-  'voici',
+  'voici_core',
   'static',
   'build'
 );
