+++ python/voici-core/setup.py
@@ -1,0 +1,2 @@
+# setup.py shim for use with applications that require it.
+__import__("setuptools").setup()
