+++ python/voici-core/voici_core/static/index.html
