+++ python/voici/voici/__init__.py
@@ -1,0 +1,1 @@
+from voici_core import *  # noqa
