+++ python/voici/voici/_version.py
@@ -1,0 +1,1 @@
+__version__ = "0.5.0"
